feat(common.shim): Add batch to shim
